Refactored HMSRoomKitNotification model to have just one id var.
